+++ bin/assets/scripts/Engine/test.lua
@@ -15,24 +15,14 @@ end
 -- Showing component Added event working
 --local ents = Manager:getEntitiesWithComponent("playerMove")
 --if ents ~= {} then ents[1]:add(ns.Component.all["topo"]()) end
-setWindowVisible("PauseWindow", true)
-setUIMouseCursor("TaharezLook/MouseArrow")
 
-createButton("PushButton", "CEGUI es facilito",vec2:new(300, 200), vec2:new(200, 50))
-setButtonFunction("PushButton","UICallback")
+setUIMouseCursor("TaharezLook/MouseArrow")
+setWindowVisible("PauseWindow", true)
+
+createButton("PushButton", "CEGUI es facilito", "TaharezLook/Button",vec2:new(300, 200), vec2:new(200, 50))
+setButtonFunction("PushButton","ShowTaiFighterUI")
+
+createButton("ExitButton", "Salir", "TaharezLook/Button",vec2:new(300, 300), vec2:new(200, 50))
+setButtonFunction("ExitButton","ExitCallback")
 
 LOG("Test.lua completed")
-
-function UICallback()
-	setWindowVisible("TaiFighterWindow", true)
-	setWindowVisible("PauseWindow", false)
-	setWindowVisible("PushButton", false)
-	setUIMouseCursorVisible(false);
-end
-
-function ShowPauseUI()
-	setWindowVisible("TaiFighterWindow", false)
-	setWindowVisible("PauseWindow", true)
-	setWindowVisible("PushButton", true)
-	setUIMouseCursorVisible(true);
-end
